Move Enum rename to its own drawer, then cleanup previous references
